requirements.txt
asgiref==3.2.3
awsebcli==3.17.0
blessed==1.17.0
boto3==1.11.9
botocore==1.14.9
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
django-debug-toolbar==2.1
django-dotenv==1.4.2
django-seed==0.1.9
django-sslserver==0.22
django-storages==1.8
djangorestframework==3.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.17.0
entrypoints==0.3
Faker==4.0.0
flake8==3.7.9
future==0.16.0
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.4
openapi-codec==1.3.2
packaging==20.1
pathspec==0.5.9
Pillow==7.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.0
pytz==2019.2
PyYAML==3.13
requests==2.20.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
s3transfer==0.3.2
semantic-version==2.5.0
simplejson==3.17.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.3
texttable==0.9.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
whitenoise==4.1.3
